Integrate single API for all endpoints
