+++ src/dbscripts/mysql.sql
@@ -15,7 +15,7 @@ CREATE TABLE "DP_ORGANIZATION_ASSETS" (
 	"FILE_TYPE" VARCHAR(256) NOT NULL,
 	"FILE_PATH" VARCHAR(1000) NOT NULL,
 	"ORG_ID" VARCHAR(256) NOT NULL,
-	FOREIGN KEY("ORG_ID") REFERENCES "DP_ORGANIZATION"("ORG_ID"),
+	FOREIGN KEY("ORG_ID") REFERENCES "DP_ORGANIZATION"("ORG_ID") ON DELETE CASCADE,
 	UNIQUE ("FILE_NAME", "FILE_TYPE", "FILE_PATH", "ORG_ID"),
 	PRIMARY KEY("ASSERT_ID")
 );
